#3875 清理重复测试文件和空壳占位测试
- 删除 14 个重复 *_pytest.py 文件（保留更完善的版本）
- 删除 6 个仅 assert True 的空壳集成测试
- 删除 integration/database/test_demo.py 空壳
- 重命名无原文件的 _pytest.py 为正式测试文件
