--- tests/integration/data/crud/test_kafka_crud_integration.py
@@ -1,26 +1,0 @@
-"""
-KafkaCRUD 集成测试
-
-注意：需要运行中的 Kafka 服务，默认跳过。
-KafkaCRUD 用于 Kafka 消息队列的 CRUD 操作，
-依赖外部 Kafka 基础设施，无法在标准集成测试环境中运行。
-"""
-import pytest
-import sys
-from pathlib import Path
-
-# 添加项目路径
-project_root = Path(__file__).parent.parent.parent.parent
-_path = str(project_root / "src")
-if _path not in sys.path:
-    sys.path.insert(0, _path)
-
-
-@pytest.mark.integration
-@pytest.mark.skip(reason="需要运行中的 Kafka 服务，无法在标准测试环境中执行")
-class TestKafkaCRUDIntegration:
-    """KafkaCRUD 集成测试（需要 Kafka 服务）"""
-
-    def test_placeholder(self):
-        """占位测试 - 需要配置 Kafka 后实现"""
-        assert True
