--- tests/unit/lab/test_sizer_base_pytest.py
@@ -1,222 +1,0 @@
-"""
-性能: 219MB RSS, 1.91s, 19 tests [PASS]
-Sizer基础测试 - 使用Pytest最佳实践重构。
-
-测试BaseSizer、FixedSizer、ATRSizer等仓位管理类的功能。
-"""
-
-import pytest
-from decimal import Decimal
-from typing import Dict, Any
-from unittest.mock import Mock, patch
-
-try:
-    from ginkgo.trading.sizers import BaseSizer, FixedSizer, ATRSizer
-    from ginkgo.entities.signal import Signal
-    from ginkgo.enums import DIRECTION_TYPES
-    SIZER_AVAILABLE = True
-except ImportError:
-    BaseSizer = None
-    FixedSizer = None
-    ATRSizer = None
-    Signal = None
-    DIRECTION_TYPES = None
-    SIZER_AVAILABLE = False
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestBaseSizerConstruction:
-    """测试BaseSizer类的构造和初始化."""
-
-    @pytest.mark.unit
-    def test_basesizer_init(self):
-        """测试BaseSizer默认初始化."""
-        if BaseSizer is None:
-            pytest.skip("BaseSizer not available")
-
-        s = BaseSizer()
-
-        # 验证基本属性
-        assert s is not None
-        assert callable(getattr(s, 'cal', None))
-
-    @pytest.mark.unit
-    def test_basesizer_portfolio_id_initial_state(self):
-        """测试BaseSizer初始化时portfolio_id状态."""
-        if BaseSizer is None:
-            pytest.skip("BaseSizer not available")
-
-        s = BaseSizer()
-
-        # 初始时portfolio_id应该为None（来自ContextMixin）
-        assert s.portfolio_id is None
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestBaseSizerProperties:
-    """测试BaseSizer类的属性管理."""
-
-    @pytest.mark.unit
-    def test_basesizer_has_cal_method(self):
-        """测试BaseSizer的cal方法存在."""
-        if BaseSizer is None:
-            pytest.skip("BaseSizer not available")
-
-        s = BaseSizer()
-
-        assert callable(getattr(s, 'cal', None))
-
-        # cal返回None (默认实现)
-        result = s.cal({}, None)
-        assert result is None
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestBaseSizerBusinessLogic:
-    """测试BaseSizer类的业务逻辑."""
-
-    @pytest.mark.unit
-    def test_basesizer_cal_default_returns_none(self):
-        """测试BaseSizer的cal方法默认返回None."""
-        if BaseSizer is None:
-            pytest.skip("BaseSizer not available")
-
-        s = BaseSizer()
-
-        # 默认实现返回None
-        result = s.cal({"positions": {}, "cash": 100000}, None)
-        assert result is None
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestFixedSizerConstruction:
-    """测试FixedSizer类的构造和初始化."""
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("fixed_volume", [100, 500, 1000, 5000])
-    def test_fixedsizer_init(self, fixed_volume):
-        """测试FixedSizer初始化."""
-        if FixedSizer is None:
-            pytest.skip("FixedSizer not available")
-
-        s = FixedSizer(volume=str(fixed_volume))
-
-        # 验证基本属性
-        assert s is not None
-
-    @pytest.mark.unit
-    def test_fixedsizer_default_initialization(self):
-        """测试FixedSizer默认初始化."""
-        if FixedSizer is None:
-            pytest.skip("FixedSizer not available")
-
-        s = FixedSizer()
-
-        # 验证默认初始化
-        assert s is not None
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestFixedSizerCalculation:
-    """测试FixedSizer的计算逻辑."""
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("fixed_volume,expected_order_volume", [
-        (100, 100),
-        (500, 500),
-        (1000, 1000),
-    ])
-    def test_fixedsizer_cal_fixed_volume(self, fixed_volume, expected_order_volume):
-        """测试FixedSizer固定数量属性."""
-        if FixedSizer is None:
-            pytest.skip("FixedSizer not available")
-
-        s = FixedSizer(volume=str(fixed_volume))
-
-        # 验证cal方法存在
-        assert callable(getattr(s, 'cal', None))
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestATRSizerConstruction:
-    """测试ATRSizer类的构造和初始化."""
-
-    @pytest.mark.unit
-    def test_atrsizer_init(self):
-        """测试ATRSizer初始化."""
-        if ATRSizer is None:
-            pytest.skip("ATRSizer not available")
-
-        s = ATRSizer()
-
-        # 验证基本属性
-        assert s is not None
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestATRSizerCalculation:
-    """测试ATRSizer的计算逻辑."""
-
-    @pytest.mark.unit
-    def test_atrsizer_cal_method_exists(self):
-        """测试ATRSizer的cal方法存在."""
-        if ATRSizer is None:
-            pytest.skip("ATRSizer not available")
-
-        s = ATRSizer()
-
-        # 验证cal方法存在
-        assert callable(getattr(s, 'cal', None))
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestSizerIntegration:
-    """测试Sizer的集成功能."""
-
-    @pytest.mark.integration
-    @pytest.mark.parametrize("volume_param", [1000, 500])
-    def test_sizer_construction(self, volume_param):
-        """测试Sizer构造."""
-        if FixedSizer is None:
-            pytest.skip("Required sizer classes not available")
-
-        s = FixedSizer(volume=str(volume_param))
-
-        # 验证构造成功
-        assert s is not None
-        assert callable(getattr(s, 'cal', None))
-
-
-@pytest.mark.lab
-@pytest.mark.sizer
-class TestSizerValidation:
-    """测试Sizer的验证功能."""
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("invalid_volume", [
-        -100,
-        0,
-        -1,
-    ])
-    def test_fixedsizer_invalid_volume(self, invalid_volume):
-        """测试FixedSizer无效数量."""
-        if FixedSizer is None:
-            pytest.skip("FixedSizer not available")
-
-        # 尝试创建无效数量的Sizer
-        # FixedSizer constructor accepts volume as string and converts to int
-        try:
-            s = FixedSizer(volume=str(invalid_volume))
-            # 如果没有验证，那么创建会成功
-            assert s is not None
-        except (ValueError, AttributeError):
-            # 如果有验证，抛出异常是预期的
-            pass
