+++ tests/unit/notifiers/test_beep.py
