+++ tests/unit/notifiers/test_mail.py
